create a shared uptimeAPI
